Changes to simplify handling the configutaion file

+++ omero_search_client/configuration/config.py
@@ -17,11 +17,17 @@ class omero_search_client_app_config (object):
     SECRET_KEY= "sdljhfdkfgsdvbflfvsdfafgdf"
     home_folder = os.path.expanduser('~')
     OMERO_SEARCH_CLIENT_INSTANCE_CONFIG = os.path.join(home_folder, '.omero_search_client.yml')
+    DEPLOYED_INSTANCE_CONFIG=r"/etc/searchengineclient/.omero_search_client.yml"
     if not os.path.isfile(OMERO_SEARCH_CLIENT_INSTANCE_CONFIG):
-        LOCAL_CONFIG_FILE = os.path.join(os.path.abspath(os.path.dirname(__file__)),
-                                         'omero_search_client.yml')
-        copyfile(LOCAL_CONFIG_FILE, OMERO_SEARCH_CLIENT_INSTANCE_CONFIG)
-        #print (LOCAL_CONFIG_FILE, OMERO_SEARCH_CLIENT_INSTANCE_CONFIG)
+        # Check if the configuration file exists in the docker deployed folder
+        # if not, it will assume it is either development environment or deploying using other methods
+        if os.path.isfile(DEPLOYED_INSTANCE_CONFIG):
+            OMERO_SEARCH_CLIENT_INSTANCE_CONFIG=DEPLOYED_INSTANCE_CONFIG
+        else:
+            LOCAL_CONFIG_FILE = os.path.join(os.path.abspath(os.path.dirname(__file__)),
+                                             'omero_search_client.yml')
+            copyfile(LOCAL_CONFIG_FILE, OMERO_SEARCH_CLIENT_INSTANCE_CONFIG)
+            #print (LOCAL_CONFIG_FILE, OMERO_SEARCH_CLIENT_INSTANCE_CONFIG)
 
 class development_app_config(omero_search_client_app_config):
     DEBUG = False
